requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.7
apscheduler==3.10.4
selenium==4.21.0
requests==2.32.2
python-multipart==0.0.9
websockets==12.0
